[clean] split file, some norm
